+++ spec/models/contact/project_spec.rb
@@ -1,44 +1,7 @@
 require "rails_helper"
-
-RSpec.describe Contact::Project, type: :model do
-  describe "Columns" do
-    it { is_expected.to have_db_column(:name).of_type :string }
-    it { is_expected.to have_db_column(:title).of_type :string }
-    it { is_expected.to have_db_column(:email).of_type :string }
-    it { is_expected.to have_db_column(:phone).of_type :string }
-    it { is_expected.to have_db_column(:organisation_name).of_type :string }
-    it { is_expected.to have_db_column(:type).of_type :string }
-  end
-
+RSpec.describe Contact::Project do
   describe "Relationships" do
-    it { is_expected.to belong_to(:project).optional }
-  end
-
-  describe "Validations" do
-    it { should validate_presence_of(:category) }
-    it { should validate_presence_of(:name) }
-    it { should validate_presence_of(:title) }
-
-    describe "#email" do
-      it { is_expected.to allow_value("test@example.com").for(:email) }
-      it { is_expected.to_not allow_value("notavalidemail").for(:email) }
-    end
-  end
-
-  describe "scopes" do
-    describe "by_name" do
-      it "orders by name" do
-        mock_successful_api_responses(urn: any_args, ukprn: any_args)
-        project = create(:conversion_project)
-        first_solicitor_contact = create(:project_contact, category: :solicitor, name: "B solicitor", project: project)
-        second_solicitor_contact = create(:project_contact, category: :solicitor, name: "A solicitor", project: project)
-
-        ordered_contacts = project.contacts.by_name
-
-        expect(ordered_contacts.first).to eq second_solicitor_contact
-        expect(ordered_contacts.last).to eq first_solicitor_contact
-      end
-    end
+    it { is_expected.to belong_to(:project) }
   end
 
   describe "#establishment_main_contact" do
@@ -59,11 +22,6 @@ RSpec.describe Contact::Project, type: :model do
       project = create(:conversion_project, establishment_main_contact_id: SecureRandom.uuid)
       contact = create(:project_contact, project: project)
 
-      expect(contact.establishment_main_contact).to be false
-    end
-
-    it "returns false if the contact does not have a project" do
-      contact = create(:project_contact, project_id: nil)
       expect(contact.establishment_main_contact).to be false
     end
   end
@@ -88,11 +46,6 @@ RSpec.describe Contact::Project, type: :model do
 
       expect(contact.incoming_trust_main_contact).to be false
     end
-
-    it "returns false if the contact does not have a project" do
-      contact = create(:project_contact, project_id: nil)
-      expect(contact.incoming_trust_main_contact).to be false
-    end
   end
 
   describe "#outgoing_trust_main_contact" do
@@ -113,11 +66,6 @@ RSpec.describe Contact::Project, type: :model do
       project = create(:conversion_project, outgoing_trust_main_contact_id: SecureRandom.uuid)
       contact = create(:project_contact, project: project)
 
-      expect(contact.outgoing_trust_main_contact).to be false
-    end
-
-    it "returns false if the contact does not have a project" do
-      contact = create(:project_contact, project_id: nil)
       expect(contact.outgoing_trust_main_contact).to be false
     end
   end